Give rules their own module and use lists
