+++ Modules/rat_command.py
@@ -13,12 +13,12 @@ This module is built on top of the Pydle system.
 """
 
 import logging
-import re
-from typing import Callable, Dict, Pattern, NamedTuple, Optional, List, Tuple
+from typing import Callable, List, Tuple
 
 from pydle import BasicClient
 
 from Modules.context import Context
+from Modules.rules import get_rule, clear_rules
 from Modules.user import User
 from config import config
 
@@ -48,8 +48,6 @@ class NameCollisionException(CommandException):
 
 
 _registered_commands = {}
-_rules: Dict[Pattern, "_RuleTuple"] = {}
-_prefixless_rules: Dict[Pattern, "_RuleTuple"] = {}
 
 # character/s that must prefix a message for it to be parsed as a command.
 prefix = config['commands']['prefix']
@@ -82,14 +80,14 @@ async def trigger(message: str, sender: str, channel: str):
             log.debug(f"Regular command {words[0]} invoked.")
         else:
             # Might be a regular rule
-            command_fun, extra_args = _get_rule(words, words_eol, _rules)
+            command_fun, extra_args = get_rule(words, words_eol, prefixless=False)
             if command_fun:
                 log.debug(f"Rule {getattr(command_fun, '__name__', '')} matching {words[0]} found.")
             else:
                 log.warning(f"Could not find command or rule for {prefix}{words[0]}.")
     else:
         # Might still be a prefixless rule
-        command_fun, extra_args = _get_rule(words, words_eol, _prefixless_rules)
+        command_fun, extra_args = get_rule(words, words_eol, prefixless=True)
         if command_fun:
             log.debug(f"Prefixless rule {getattr(command_fun, '__name__', '')} matching {words[0]} "
                       f"found.")
@@ -100,36 +98,6 @@ async def trigger(message: str, sender: str, channel: str):
         return await command_fun(context, *extra_args)
     else:
         log.debug(f"Ignoring message '{message}'. Not a command or rule.")
-
-
-def _get_rule(words: List[str], words_eol: List[str],
-              from_dict: Dict[Pattern, "_RuleTuple"]) -> Tuple[Optional[Callable], tuple]:
-    """
-    Attempt to find a rule in the given dict of patterns and rules.
-
-    Args:
-        words: Words of the message.
-        words_eol: Words of the message, but each including everything to the end of it.
-        from_dict: Dict mapping patterns to their rules.
-
-    Returns:
-        (Callable or None, tuple):
-            2-tuple of the command function and the extra args that it should
-            be called with.
-    """
-    for pattern, (fun, full_message, pass_match) in from_dict.items():
-        if full_message:
-            match = pattern.match(words_eol[0])
-        else:
-            match = pattern.match(words[0])
-
-        if match is not None:
-            if pass_match:
-                return fun, (match,)
-            else:
-                return fun, ()
-    else:
-        return None, ()
 
 
 def _split_message(string: str) -> Tuple[List[str], List[str]]:
@@ -200,9 +168,9 @@ def _flush() -> None:
     Flushes registered commands
     Probably useless outside testing...
     """
-    global _registered_commands, _rules  # again this feels ugly but they are module-level now...
+    global _registered_commands  # again this feels ugly but they are module-level now...
     _registered_commands = {}
-    _rules = {}
+    clear_rules()
 
 
 def command(*aliases):
@@ -231,49 +199,3 @@ def command(*aliases):
 
         return func
     return real_decorator
-
-
-_RuleTuple = NamedTuple("_RuleTuple", underlying=Callable, full_message=bool, pass_match=bool)
-
-
-def rule(regex: str, *, case_sensitive: bool=False, full_message: bool=False,
-         pass_match: bool=False, prefixless: bool=False):
-    """
-    Decorator to have the underlying coroutine be called when two conditions apply:
-    1. No conventional command was found for the incoming message.
-    2. The command matches the here provided regular expression.
-
-    Arguments:
-        regex (str):
-            Regular expression to match the command.
-        case_sensitive (bool):
-            Whether to match case-sensitively (using the re.IGNORECASE flag).
-        full_message (bool):
-            If this is True, will try to match against the full message. Otherwise,
-            only the word will be matched against.
-        pass_match (bool):
-            If this is True, the match object will be passed as an argument toward the command
-            function.
-        prefixless (bool):
-            If this is True, the rule can match whether or not the prefix is present. Otherwise, it
-            acts similarly to a command, only being considered if the message starts with our
-            prefix. The prefix itself should not be part of the regex for non-prefixless rules.
-
-    Please note that *regex* can match anywhere within the string, it need not match the entire
-    string. If you wish to change this behaviour, use '^' and '$' in your regex.
-    """
-    def decorator(coro: Callable):
-        if case_sensitive:
-            pattern = re.compile(regex)
-        else:
-            pattern = re.compile(regex, re.IGNORECASE)
-
-        if prefixless:
-            _prefixless_rules[pattern] = _RuleTuple(coro, full_message, pass_match)
-        else:
-            _rules[pattern] = _RuleTuple(coro, full_message, pass_match)
-
-        log.info(f"New rule matching '{regex}' case-{'' if case_sensitive else 'in'}sensitively was"
-                 f" created.")
-        return coro
-    return decorator
